requirements.txt
requests
elasticsearch
python-dotenv
urllib3
ollama
argparse

